+++ src/prelude.rs
@@ -155,16 +155,17 @@ Smalltalk, LISP and Perl.
   - [4.11](#411-symbols) - Symbols
     - [4.11.1](#4111-stdsymbolscollect) - std:symbols:collect
   - [4.12](#412-pairs-pa-b) - Pairs `$p(a, b)`
-    - [4.12.1](#4121-cons-a-b) - cons _a_ _b_
-    - [4.12.2](#4122-pair-stringbyte-vector-operations) - Pair string/byte vector operations
-      - [4.12.2.1](#41221-pfrom-count-string-or-byte-vec) - `$p(_from_, _count_) _string-or-byte-vec_`
-      - [4.12.2.2](#41222-ppattern-replacement-string-or-byte-vec) - `$p(_pattern_, _replacement_) _string-or-byte-vec_`
-      - [4.12.2.3](#41223-psplit-pattern-max-string-or-byte-vec) - `$p(_split-pattern_, _max_) _string-or-byte-vec_`
-    - [4.12.3](#4123-pair-to-iterator) - Pair to Iterator
-      - [4.12.3.1](#41231-iter---enumerate) - Iter - Enumerate
-      - [4.12.3.2](#41232-iter---values) - Iter - Values
-      - [4.12.3.3](#41233-iter---keys) - Iter - Keys
-    - [4.12.4](#4124-ispair-value) - is_pair _value_
+    - [4.12.1](#4121-pair-operator-a--b) - Pair Operator `a => b`
+    - [4.12.2](#4122-cons-a-b) - cons _a_ _b_
+    - [4.12.3](#4123-pair-stringbyte-vector-operations) - Pair string/byte vector operations
+      - [4.12.3.1](#41231-pfrom-count-string-or-byte-vec) - `$p(_from_, _count_) _string-or-byte-vec_`
+      - [4.12.3.2](#41232-ppattern-replacement-string-or-byte-vec) - `$p(_pattern_, _replacement_) _string-or-byte-vec_`
+      - [4.12.3.3](#41233-psplit-pattern-max-string-or-byte-vec) - `$p(_split-pattern_, _max_) _string-or-byte-vec_`
+    - [4.12.4](#4124-pair-to-iterator) - Pair to Iterator
+      - [4.12.4.1](#41241-iter---enumerate) - Iter - Enumerate
+      - [4.12.4.2](#41242-iter---values) - Iter - Values
+      - [4.12.4.3](#41243-iter---keys) - Iter - Keys
+    - [4.12.5](#4125-ispair-value) - is_pair _value_
   - [4.13](#413-vectors-or-lists) - Vectors (or Lists)
     - [4.13.1](#4131-stdpush-vector-item) - std:push _vector_ _item_
     - [4.13.2](#4132-stdpop-vector) - std:pop _vector_
@@ -207,6 +208,10 @@ Smalltalk, LISP and Perl.
   - [7.1](#71-control-flow) - Control Flow
     - [7.1.1](#711-while-predicate-body) - while _predicate_ _body_
     - [7.1.2](#712-iter-var-iterable-body) - iter _var_ _iterable_ _body_
+      - [7.1.2.1](#7121-counting-loop-with-iter) - Counting loop with _iter_
+      - [7.1.2.2](#7122-vector-iteration-with-iter) - Vector iteration with _iter_
+      - [7.1.2.3](#7123-map-iteration-with-iter) - Map iteration with _iter_
+      - [7.1.2.4](#7124-closures-and-iter-iter-i-) - Closures and _iter_ `iter $&&i ...`
     - [7.1.3](#713-range-start-end-step-fun) - range _start_ _end_ _step_ _fun_
     - [7.1.4](#714-break-value) - break _value_
     - [7.1.5](#715-next) - next
@@ -246,6 +251,10 @@ Smalltalk, LISP and Perl.
     - [8.3.3](#833--op-a-op-b) - &| _op-a_ _op-b_
     - [8.3.4](#834--op-a-op-b) - << _op-a_ _op-b_
     - [8.3.5](#835--op-a-op-b) - >> _op-a_ _op-b_
+  - [8.4](#84-syntactic-operators) - Syntactic Operators
+    - [8.4.1](#841-pair-constructor-a--b) - Pair Constructor `a => b`
+    - [8.4.2](#842-forward-argument-pipe-arg--fun) - Forward Argument Pipe `arg &> fun`
+    - [8.4.3](#843-reverse-argument-pipe-fun--arg) - Reverse Argument Pipe `fun <& arg`
 - [9](#9-data-structure-selectors-and-string-patternsregex) - Data Structure Selectors and String Patterns/Regex
   - [9.1](#91-data-structure-selectors) - Data Structure Selectors
   - [9.2](#92-string-patterns-regex) - String Patterns (Regex)
@@ -2543,6 +2552,8 @@ std:assert std:ref_id[a] == id_a;
 std:assert std:ref_id[v.0] == id_a;
 ```
 
+#### <a name="4121-pair-operator-a--b"></a>4.12.1 - Pair Operator `a => b`
+
 You can construct pairs also using the `=>` operator, which
 has the lowest precedence and is right associative:
 
@@ -2555,7 +2566,7 @@ std:assert_eq b $p(2, $p(3, 4));
 std:assert_eq b 2 => 3 => 4;
 ```
 
-#### <a name="4121-cons-a-b"></a>4.12.1 - cons _a_ _b_
+#### <a name="4122-cons-a-b"></a>4.12.2 - cons _a_ _b_
 
 Creates a new pair from the values _a_ and _b_.
 
@@ -2565,12 +2576,12 @@ Creates a new pair from the values _a_ and _b_.
 std:assert_eq p $p(3, 4);
 ```
 
-#### <a name="4122-pair-stringbyte-vector-operations"></a>4.12.2 - Pair string/byte vector operations
+#### <a name="4123-pair-stringbyte-vector-operations"></a>4.12.3 - Pair string/byte vector operations
 
 If you call a pair with a string or byte vector as argument, there are some
 operations that can be done:
 
-##### <a name="41221-pfrom-count-string-or-byte-vec"></a>4.12.2.1 - `$p(_from_, _count_) _string-or-byte-vec_`
+##### <a name="41231-pfrom-count-string-or-byte-vec"></a>4.12.3.1 - `$p(_from_, _count_) _string-or-byte-vec_`
 
 Returns a substring starting at _from_ with the length _count_.
 
@@ -2584,7 +2595,7 @@ The same works for byte vectors:
 std:assert_eq ($p(2, 4) $b"abcdefgh") $b"cdef";
 ```
 
-##### <a name="41222-ppattern-replacement-string-or-byte-vec"></a>4.12.2.2 - `$p(_pattern_, _replacement_) _string-or-byte-vec_`
+##### <a name="41232-ppattern-replacement-string-or-byte-vec"></a>4.12.3.2 - `$p(_pattern_, _replacement_) _string-or-byte-vec_`
 
 Replaces all _pattern_ occurences in _string_ by _replacement_.
 
@@ -2598,7 +2609,7 @@ The same works for byte vectors:
 std:assert_eq ($p($b";", $b"_") $b"A;B;D;EFG;HI") $b"A_B_D_EFG_HI";
 ```
 
-##### <a name="41223-psplit-pattern-max-string-or-byte-vec"></a>4.12.2.3 - `$p(_split-pattern_, _max_) _string-or-byte-vec_`
+##### <a name="41233-psplit-pattern-max-string-or-byte-vec"></a>4.12.3.3 - `$p(_split-pattern_, _max_) _string-or-byte-vec_`
 
 Splits _string_ at _split-pattern_ a _max_ number of times.
 If _max_ is 0, it is split completely.
@@ -2615,14 +2626,28 @@ The same works for byte vectors:
 std:assert_eq str[$p($b";", 0) $b"A;B;D;EFG;HI"] ~ str $[$b"A", $b"B", $b"D", $b"EFG", $b"HI"];
 ```
 
-#### <a name="4123-pair-to-iterator"></a>4.12.3 - Pair to Iterator
+#### <a name="4124-pair-to-iterator"></a>4.12.4 - Pair to Iterator
 
 Pairs play a special role if you make an iterator from it.
 It can be used to create a specialized iterator that only
 iterates over keys or values of a map. Or that enumerates
 a vector or map.
 
-##### <a name="41231-iter---enumerate"></a>4.12.3.1 - Iter - Enumerate
+##### - Iter - Range
+
+`$iter $p(0, 10)` is the same as `$iter $i(0, 10)` and will construct an
+iterator that iterates from `0` to `9` (inclusive).
+Because of the pair operator `a => b` we can nicely write a counting loop like this:
+
+```wlambda
+!sum = $@i
+    iter i 0 => 10 {
+        $+ i;
+    };
+std:assert_eq sum 45;
+```
+
+##### <a name="41241-iter---enumerate"></a>4.12.4.1 - Iter - Enumerate
 
 If the first value of the pair is `:enumerate`
 it will enumerate entries in a map or values in a vector.
@@ -2649,7 +2674,7 @@ iter i $p(:enumerate, ${a = 10, b = 20})
 std:assert_eq (str v) (str $[0, 1]);
 ```
 
-##### <a name="41232-iter---values"></a>4.12.3.2 - Iter - Values
+##### <a name="41242-iter---values"></a>4.12.4.2 - Iter - Values
 
 This is useful for iterating over the values in a map in an undefined order:
 
@@ -2661,7 +2686,7 @@ This is useful for iterating over the values in a map in an undefined order:
 std:assert_eq sum 63;
 ```
 
-##### <a name="41233-iter---keys"></a>4.12.3.3 - Iter - Keys
+##### <a name="41243-iter---keys"></a>4.12.4.3 - Iter - Keys
 
 You can also iterate over map keys in an undefined order:
 
@@ -2673,7 +2698,7 @@ You can also iterate over map keys in an undefined order:
 std:assert_eq sum 60;
 ```
 
-#### <a name="4124-ispair-value"></a>4.12.4 - is_pair _value_
+#### <a name="4125-ispair-value"></a>4.12.5 - is_pair _value_
 
 Checks if _value_ is a pair.
 
@@ -3128,6 +3153,7 @@ Here is a table of the behaviour of iterators created from WLambda data.
 | `$p(:enumerate, vector)`  | Returns integers in the range of `0` to `len vector`. |
 | `$p(:values, map)`  | Returns the values of the _map_ in undefined order. |
 | `$p(:keys, map)`  | Returns the keys of the _map_ in undefined order. |
+| `$p(int_a, int_b)` | The same as `$i(a, b)`. This makes it possible to write `$iter 0 => 10`. |
 | `$p(iterator_a, iterator_b)` | Returns a zip operation of the elements returned by the iterator_a and iterator_b until one of both returns `$o()`. |
 | `$p(iterator, x)` | Returns a zip operation of the elements returned by the iterator and the newly created iterator`$iter x`. |
 
@@ -3640,6 +3666,28 @@ See further below for an example.
 Like usual, the control flow manipulators `next` and `break` also work
 for this kind of loop.
 
+##### <a name="7121-counting-loop-with-iter"></a>7.1.2.1 - Counting loop with _iter_
+
+Here is an example how to iterate over a range from 1 to 9 and
+collect the sum of those integers using an accumulator:
+
+```wlambda
+!sum = $@int iter i $i(1,10) ~ $+ i;
+
+std:assert_eq sum 45;
+```
+
+Because `$iter $p(1, 10)` is the same as `$iter $i(1, 10)` and because
+there is the pair constructor operator `a => b`, the above can also be written as:
+
+```wlambda
+!sum = $@int iter i 1 => 10 ~ $+ i;
+
+std:assert_eq sum 45;
+```
+
+##### <a name="7122-vector-iteration-with-iter"></a>7.1.2.2 - Vector iteration with _iter_
+
 Here is a simple example of how to iterate over all items of a vector
 in order:
 
@@ -3690,14 +3738,7 @@ iter i $i(0, len v) {
 std:assert_eq sum 15;
 ```
 
-Here is an example how to iterate over a range from 1 to 9 and
-collect the sum of those integers using an accumulator:
-
-```wlambda
-!sum = $@int iter i $i(1,10) ~ $+ i;
-
-std:assert_eq sum 45;
-```
+##### <a name="7123-map-iteration-with-iter"></a>7.1.2.3 - Map iteration with _iter_
 
 Iteration over a map is also easy and concise. The map entry
 will be represented using a pair value `$p(value, key)`.
@@ -3742,6 +3783,8 @@ iter k $p(:keys, m) {
 
 std:assert_eq sum 30;
 ```
+
+##### <a name="7124-closures-and-iter-iter-i-"></a>7.1.2.4 - Closures and _iter_ `iter $&&i ...`
 
 If you need a new variable for capturing it in a closure on each
 iteration you can use the special `iter $&&i ...` syntax:
@@ -4386,6 +4429,59 @@ std:assert (0b0011 >> 2)      == 0b0;
 std:assert (0b1100 >> 2)      == 0b11;
 std:assert (`>>` 0b1011000 3) == 0b1011
 ```
+
+### <a name="84-syntactic-operators"></a>8.4 - Syntactic Operators
+
+This section has some very special syntactic operators which are handled by the
+parser. They are for constructing pairs and calling functions.
+
+#### <a name="841-pair-constructor-a--b"></a>8.4.1 - Pair Constructor `a => b`
+
+Writing `a => b` operator is the same as writing `$p(a, b)`.
+However, the precedence of the `=>` operator is the lowest, so writing this
+is possible:
+
+```wlambda
+!p = 1 + 2 => 3 + 4;
+
+std:assert_eq p $p(3, 7);
+```
+
+#### <a name="842-forward-argument-pipe-arg--fun"></a>8.4.2 - Forward Argument Pipe `arg &> fun`
+
+This operator has the highest precedence over all other operators
+and is used to be able to write this:
+
+```wlambda
+? "foob" &> $r/f(^*)b/ {
+    std:assert_eq $\.1 "oo";
+} {
+    std:assert $false;
+}
+```
+
+That means `f a &> b` is equivalent to writing `f[b[a]]` or `(f (b a))`.
+Chaining multiple is also possible and left associative: `a &> b &> c` is `(c (b a))`.
+You can see it as piping operation:
+
+```wlambda
+!r = "ABC" &> std:str:to_lowercase &> \std:str:pad_start 10 "0" _;
+
+std:assert_eq r "0000000abc";
+```
+
+#### <a name="843-reverse-argument-pipe-fun--arg"></a>8.4.3 - Reverse Argument Pipe `fun <& arg`
+
+Like the `&>` operator this operator, but it has a lower precedence (does not bind
+as strongly as `&>`) and is right associative. That means you can write this:
+
+```wlambda
+!r = (\std:str:pad_start 10 "0" _) <& std:str:to_lowercase <& "ABC";
+
+std:assert_eq r "0000000abc";
+```
+
+That means, writing `f <& a <& x` becomes `f[a[x]]` or `(f (a x))`.
 
 ## <a name="9-data-structure-selectors-and-string-patternsregex"></a>9 - Data Structure Selectors and String Patterns/Regex