+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 * Feature: `iter $&&i ...` syntax implemented, for creating new references
 for each iteration for easier closure generation.
 * Feature: Implemented pair operator `a => b` which does the same as `$p(a, b)`.
+* Feature: `$iter $p(0, 10)` does now the same as `$iter $i(0, 10)`, allowing
+nice composition with the pair operator: `$iter 0 => 10`.
+* Feature: Implemented argument function call operators `a &> f` and `f <& a`,
+so you can write regular expressions like `? "foo" &> $r/a*b/ { ... }`
+without parenthesis.
+* Incompatible Change: `a + b ~ c` is no longer `a + (b ~ c)` but `(a + b) ~ c`
+to let `iter i 0 => 10 ~ std:displayln i` work properly.
 
 0.5.0 (2020-05-21)
 ==================